backend/requirements.txt
arrow==1.0.3
asgiref==3.3.1
binaryornot==0.4.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cookiecutter==1.7.2
Django==3.1.7
django-cors-headers==3.7.0
django-summernote==0.8.11.6
djangorestframework==3.12.2
idna==2.10
Jinja2==2.11.3
jinja2-time==0.2.0
MarkupSafe==1.1.1
Pillow==8.1.2
pkg-resources==0.0.0
poyo==0.5.0
python-dateutil==2.8.1
python-slugify==4.0.1
python-webpack-boilerplate==0.0.2
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
